Consolidate auth types into transport/auth and flatten internal/auth into internal
Move Principal, AuthBackend interface, context helpers, and API key utilities
into transport/auth to eliminate the daemon adapter layer and duplicated types
in transport/http.go. Flatten internal/auth/ (backend.go, cache.go) into the
internal package as AuthBackend/AuthCache, removing the confusing auth vs tauth
import alias situation.
